requirements.txt
# ============================================================
# Dependencies: every component earns its spot here
# ============================================================
# Every package here does something specific and necessary.
# If you can't explain why it's here in one sentence, it
# shouldn't be here.
# ============================================================

# Core data wrangling: the foundation everything sits on
pandas>=2.0.0
numpy>=1.24.0

# GCP Services: the managed service backbone
google-genai>=1.0.0                # Gemini API (generative language)
google-cloud-language>=2.11.0      # NL API for entity extraction & sentiment
google-cloud-storage>=2.14.0       # Cloud Storage for raw data
google-cloud-bigquery>=3.14.0      # BigQuery for production document queries
google-cloud-firestore>=2.14.0     # Firestore for agent session memory

# NLP Baselines: because you should always have something to compare against
spacy>=3.7.0                       # Local NER baseline (fast, free, offline)
presidio-analyzer>=2.2.0           # Microsoft Presidio PII-focused NER baseline
sumy>=0.11.0                       # Extractive summarisation baseline
nltk>=3.8.0                        # Tokenisation and text stats

# Data & Evaluation: you can't improve what you don't measure
datasets>=2.16.0                   # HuggingFace datasets (CNN/DailyMail loader)
rouge-score>=0.1.2                 # ROUGE metrics for summarisation quality
bert-score>=0.3.13                 # BERTScore semantic similarity

# Visualisation: making the EDA notebooks look good
matplotlib>=3.8.0
seaborn>=0.13.0
wordcloud>=1.9.0

# Utilities: the little things that keep life sane
pyyaml>=6.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0
fpdf2>=2.7.0                       # PDF report generation
